+++ backend/apps/users/api/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path, include
+
+
+urlpatterns = [
+    path("v1/", include("apps.users.api.v1.urls")),
+]
