--- backend/apps/users/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path, include
-
-
-urlpatterns = [
-    path("registration/", include("apps.users.registration.urls"))
-]
